backend/requirements.txt
alembic
asgi_lifespan
argon2_cffi
asyncio
asyncpg
bcrypt
black
colorama
email-validator
fastapi
greenlet
httpx
isort
loguru
mypy
passlib
pathlib
pre-commit
python-dotenv
python-decouple
python-jose
python-slugify
pytest
pytest-asyncio
pytest-cov
pytest-xdist
SQLAlchemy
trio
uvicorn
pydantic-settings
psycopg2-binary
asyncpg
minio